assignment-3/python/requirements.txt
requests
pytest
jsonschema
pact-python
responses
